% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{CNDRGAll}
\alias{CNDRGAll}
\title{CNDRGAll}
\usage{
CNDRGAll(data, adrgruleslst = CNDRGRulesLst20170619, trans = T)
}
\description{
CNDRGAll takes fn_Trans fn_excl fn_cc and CNDRGGrouper together for grouping
}

