% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{CNDRGGrouper}
\alias{CNDRGGrouper}
\title{CNDRGGrouper}
\usage{
CNDRGGrouper(data, ADRGRulesLst, groupVars = c("age", "bornWt", "statusOut",
  "LOS", "PDX", "PPROC", "ADX", "APROC", "CC"), weight = c(1, 2, 4, 8, 16, 32,
  64, 128, 256))
}
\description{
CNDRGGrouper takes data(which is transformed by dataTrans into standard data) and return ADRG results containing MDC code, MDC name, ADRG code, ADRG name.
}

