% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{CNDRGGrouperAll}
\alias{CNDRGGrouperAll}
\title{CNDRGGrouperAll}
\usage{
CNDRGGrouperAll(data, adrgruleslst = CNDRGRulesLst20170619)
}
\description{
CNDRGGrouper all takes fn_Trans fn_excl fn_cc and CNDRGGrouper together for grouping
}

