% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{fn_ADRGRulesLst}
\alias{fn_ADRGRulesLst}
\title{fn_ADRGRulesLst}
\usage{
fn_ADRGRulesLst(adrgrules, rulevars = c("age", "bornWt", "statusOut", "LOS",
  "PDX", "PPROC", "CC"))
}
\description{
transform ADRGRulesDf to ageRuleDf,bornWtRuleDf,PDXRuleDf,PPROCRuleDf objs.
}

